create updating labels

+++ MyWeather/Networking/Model/WeaherInfo.swift
@@ -23,7 +23,7 @@ struct Clouds: Codable {
 
 // MARK: - Coord
 struct Coord: Codable {
-    let lon, lat: Int?
+    let lon, lat: Double?
 }
 
 // MARK: - Main
@@ -65,3 +65,4 @@ struct Wind: Codable {
     let deg: Int?
     let gust: Double?
 }
+
